requirement.txt
Flask==2.3.2
tensorflow
numpy
mysql-connector-python
Flask-Cors
joblib
Pillow
